Fix #3

+++ HFCore/ImageHelper.cs
@@ -6,7 +6,7 @@ namespace HaltroyFramework
 {
     public static class ImageHelper
     {
-        private static float[][] _colorMatrixElements = { 
+        private static float[][] _colorMatrixElements = {
                                             new float[] {(float)0.299, (float)0.299, (float)0.299, 0, 0},
                                             new float[] {(float)0.587, (float)0.587, (float)0.587, 0, 0},
                                             new float[] {(float)0.114, (float)0.114, (float)0.114, 0, 0},
@@ -29,8 +29,8 @@ namespace HaltroyFramework
             // and the Math.Min improvement from http://stackoverflow.com/questions/6501797/resize-image-proportionally-with-maxheight-and-maxwidth-constraints
 
             // Figure out the ratio
-            double ratioX = (double)canvasSize.Width / (double)imageSize.Width;
-            double ratioY = (double)canvasSize.Height / (double)imageSize.Height;
+            double ratioX = canvasSize.Width / (double)imageSize.Width;
+            double ratioY = canvasSize.Height / (double)imageSize.Height;
 
             // use whichever multiplier is smaller
             double ratio = Math.Min(ratioX, ratioY);
